% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddComponents}
\alias{AddComponents}
\title{Add a component to a problem list object (internal)}
\usage{
AddComponents(
  ThisProblem,
  CompName,
  CompCharge,
  CompMCName = NULL,
  CompType,
  CompActCorr,
  CompSiteDens,
  CompMCR = match(CompMCName, ThisProblem$MassName)
)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{CompName}{A character vector with the name(s) of the components to be
added.}

\item{CompCharge}{An integer vector with the charge(s) of the components to
be added.}

\item{CompMCName}{A character vector with the name(s) of the mass
compartments the new components are associated with. Does not need to be
specified if `CompMCR` is specified instead.}

\item{CompType}{A character vector with the types of the new input variables.
Must be one of "MassBal", "FixedAct", "FixedConc", "DonnanHA", or
"DonnanFA".}

\item{CompActCorr}{A character vector with the activity correction method(s)
of the new components. Must be one of "None", "Debye", "Davies",
"DonnanHA", "DonnanFA", "WHAMHA", or "WHAMFA".}

\item{CompSiteDens}{A numeric vector with the binding site densities of the
new components.}

\item{CompMCR}{(optional) A character vector with the indices of the mass
compartments the new components are associated with. Only needs to be
specified if `CompMCName` is not specified.}
}
\value{
`ThisProblem`, with the added component.
}
\description{
An component should be added either as an input component or a defined
component. Both of those functions will call this function, but components
should generally not be added without specifying how to obtain numbers for
them from inputs.
}
