% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddDefComp}
\alias{AddDefComp}
\title{Add a defined component to a problem list object}
\usage{
AddDefComp(
  ThisProblem,
  DefCompName,
  DefCompFromNum = NULL,
  DefCompFromVar = NULL,
  DefCompCharge,
  DefCompMCName = NULL,
  DefCompType,
  DefCompActCorr,
  DefCompSiteDens,
  DefCompMCR = match(DefCompMCName, ThisProblem$MassName)
)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{DefCompName}{A character vector with the name(s) of the components to
be added.}

\item{DefCompFromNum}{A numeric vector with the numeric values used to derive
the component. Specify `NA` if the defined component uses a variable to
define it.}

\item{DefCompFromVar}{A character vector with the variable names used to
derive the component. Specify `NA` if the defined component uses a number
to define it.}

\item{DefCompCharge}{An integer vector with the charge(s) of the components
to be added.}

\item{DefCompMCName}{A character vector with the name(s) of the mass
compartments the new components are associated with. Does not need to be
specified if `DefCompMCR` is specified instead.}

\item{DefCompType}{A character vector with the types of the new defined
variables. Must be one of "MassBal", "FixedAct", "FixedConc", "DonnanHA",
or "DonnanFA".}

\item{DefCompActCorr}{A character vector with the activity correction
method(s) of the new components. Must be one of "None", "Debye", "Davies",
"DonnanHA", "DonnanFA", "WHAMHA", or "WHAMFA".}

\item{DefCompSiteDens}{A numeric vector with the binding site densities of
the new components.}

\item{DefCompMCR}{(optional) A character vector with the indices of the mass
compartments the new components are associated with. Only needs to be
specified if `DefCompMCName` is not specified.}
}
\value{
`ThisProblem`, with the added defined component.
}
\description{
Add a defined component to a problem list object
}
