% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddInComp}
\alias{AddInComp}
\title{Add an input component to a problem list object}
\usage{
AddInComp(
  ThisProblem,
  InCompName,
  InCompCharge,
  InCompMCName = NULL,
  InCompType,
  InCompActCorr,
  InCompMCR = match(InCompMCName, ThisProblem$MassName)
)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{InCompName}{A character vector with the name(s) of the components to be
added.}

\item{InCompCharge}{An integer vector with the charge(s) of the components to
be added.}

\item{InCompMCName}{A character vector with the name(s) of the mass
compartments the new components are associated with. Does not need to be
specified if `InCompMCR` is specified instead.}

\item{InCompType}{A character vector with the types of the new input
variables. Must be one of "MassBal", "FixedAct", "FixedConc", "DonnanHA",
or "DonnanFA".}

\item{InCompActCorr}{A character vector with the activity correction
method(s) of the new components. Must be one of "None", "Debye", "Davies",
"DonnanHA", "DonnanFA", "WHAMHA", or "WHAMFA".}

\item{InCompMCR}{(optional) A character vector with the indices of the mass
compartments the new components are associated with. Only needs to be
specified if `InCompMCName` is not specified.}
}
\value{
`ThisProblem`, with the added input component.
}
\description{
Add an input component to a problem list object
}
