% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddInLabs}
\alias{AddInLabs}
\title{Add an input label to a problem list object}
\usage{
AddInLabs(ThisProblem, InLabName)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{InLabName}{A character vector with the name(s) of the new input
label(s).}
}
\value{
`ThisProblem`, with the added input labels.
}
\description{
Add an input label to a problem list object
}
