% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddInVar}
\alias{AddInVar}
\title{Add an input variable to a problem list object}
\usage{
AddInVar(
  ThisProblem,
  InVarName,
  InVarMCName = NULL,
  InVarType,
  InVarMCR = match(InVarMCName, ThisProblem$MassName)
)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{InVarName}{A character vector with the name(s) of the input input
variable(s).}

\item{InVarMCName}{A character vector with the name(s) of the mass
compartments the new input variables are associated with. Does not need to
be specified if `InVarMCR` is specified instead.}

\item{InVarType}{A character vector with the types of the new input
variables. Must be one of "Temp", "pH", "WHAM-HA", "WHAM-FA", "WHAM-HAFA",
"PercHA", and "PercAFA".}

\item{InVarMCR}{(optional) A character vector with the indices of the mass
compartments the new input variables are associated with. Only needs to be
specified if `InVarMCName` is not specified.}
}
\value{
`ThisProblem`, with the added input variable. If the input variable
  is pH, an "H" component will also be added as a fixed activity component.
}
\description{
Add an input variable to a problem list object
}
