% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddMassCompartments}
\alias{AddMassCompartments}
\title{Add a mass compartment to a problem list object}
\usage{
AddMassCompartments(ThisProblem, MassName, MassAmt, MassUnit)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{MassName}{A character vector with the name(s) of the new mass
compartment(s).}

\item{MassAmt}{A numeric vector with the mass compartment amount(s).}

\item{MassUnit}{A character vector with the units for the amount(s) of the
mass compartment(s).}
}
\value{
`ThisProblem`, with the added mass compartments.
}
\description{
Add a mass compartment to a problem list object
}
