% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{AddSpecies}
\alias{AddSpecies}
\title{Add a species reaction to a problem list object}
\usage{
AddSpecies(
  ThisProblem,
  SpecName,
  SpecMCName = NULL,
  SpecActCorr,
  SpecCompNames = list(),
  SpecCompStoichs = list(),
  SpecLogK,
  SpecDeltaH,
  SpecTempKelvin,
  SpecMCR = match(SpecMCName, ThisProblem$MassName)
)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{SpecName}{A character vector with the name(s) of the species to add
formation reactions for.}

\item{SpecMCName}{A character vector with the name(s) of the mass
compartments the new species are associated with. Does not need to be
specified if `SpecMCR` is specified instead.}

\item{SpecActCorr}{A character vector with the activity correction method(s)
of the new species. Must be one of "None", "Debye", "Davies", "DonnanHA",
"DonnanFA", "WHAMHA", or "WHAMFA".}

\item{SpecCompNames}{A list where each element is a character vector of the
component names used to form each species.}

\item{SpecCompStoichs}{A list where each element is an integer vector of the
stoichiometric coefficients of each component used to form each species.}

\item{SpecLogK}{A numeric vector with the log10-transformed equilibrium
coefficients of the species formation reactions.}

\item{SpecDeltaH}{A numeric vector with the change in enthalpy of the species
formation reactions.}

\item{SpecTempKelvin}{A numeric vector with the temperatures (in Kelvin)
corresponding to `SpecDeltaH` values of the species formation reactions.}

\item{SpecMCR}{(optional) A character vector with the indices of the mass
compartments the new species are associated with. Only needs to be
specified if `SpecMCName` is not specified.}
}
\value{
`ThisProblem`, with the added species reaction(s).
}
\description{
Add a species reaction to a problem list object
}
\examples{

# Make a blank problem:
ThisProblem = BlankProblem()

# Add Water as a mass compartment
ThisProblem = AddMassCompartments(ThisProblem, "Water", 1, "L")

# Add temperature and pH variables:
ThisProblem = AddInVar(ThisProblem,
                       InVarName = c("Temperature", "pH"),
                       InVarMCName = rep("Water", 2),
                       InVarType = c("Temp", "pH"))

# Add CO3 as a component (H already added with pH):
ThisProblem = AddInComp(
  ThisProblem,
  InCompName = "CO3",
  InCompCharge = -2,
  InCompMCName = "Water",
  InCompType = "MassBal",
  InCompActCorr = "Debye"
)

# Add reactions:
# OH = -H            logK = -14
# HCO3 = H + CO3     logK = 10.329
# H2CO3 = 2*H + CO3  logK = 10.329 + 6.352 = 16.681
ThisProblem = AddSpecies(
  ThisProblem,
  SpecName = c("OH", "HCO3", "H2CO3"),
  SpecMCName = rep("Water", 3),
  SpecActCorr = rep("Debye", 3),
  SpecCompNames = list(c("H"), c("H", "CO3"), c("H", "CO3")),
  SpecCompStoichs = list(c(-1), c(1, 1), c(1, 2)),
  SpecLogK = c(-14, 10.329, 16.681),
  SpecDeltaH = c(56186.73738, -14997.55155, -24166.23162),
  SpecTempKelvin = rep(298.1514609, 3)
)

}
