% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{AdjustForToxMode}
\alias{AdjustForToxMode}
\title{Adjust the residual and error for toxicity mode}
\usage{
AdjustForToxMode(NBLMetal, BLMetalSpecs, MetalComp, CATarget, 
  SpecConc, Resid, CompError)
}
\arguments{
\item{NBLMetal}{integer, the number of biotic ligand-bound metal species}

\item{SpecConc}{numeric vector (NSpec), the concentration of chemical
species}

\item{MetalComp}{integer, the position in component vectors of the toxic
metal component}

\item{BLMetalSpecs}{integer vector, the position in the species vectors of
the metal-biotic ligand species associated with toxicity}

\item{CATarget}{numeric, the target critical accumulation value for a
toxicity run, in mol/kg}

\item{Resid}{(return value) numeric vector (NComp), the residuals =
calculated totals - known totals, modified for toxicity mode upon return}

\item{CompError}{(return value) numeric vector (NComp), the absolute error
fraction for each component in this iteration =abs(Resid / TotMoles),
modified for toxicity mode upon return}
}
\description{
Calculate a new residual and error value for the toxic metal
  component, based instead on the difference between the target and 
  calculated critical accumulations.
}
\details{
This should only be run when DoTox = TRUE. The residual for the
  `MetalComp` component is instead the sum of the `BLMetalSpecs`
  concentrations minus the `CATarget`. This, paired with a modified set of 
  derivatives for the `MetalComp` in the Jacobian matrix, results in 
  Newton-Rapshon changing the metal concentration in order to produce
  chemical conditions that would result in the critical accumulation known
  to cause a toxic effect.
}
