% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{All_NIST20170203_reactions}
\alias{All_NIST20170203_reactions}
\title{All NIST_20170203.dbs reactions}
\format{
An object of class \code{list} of length 24.
}
\usage{
All_NIST20170203_reactions
}
\description{
A large problem using the WHAM V "NIST_20170203.dbs"
  thermodynamic database. This is the thermodynamic database used in some of
  the newer Windows BLM parameter files, including the Environment and
  Climate Change Canada (ECCC) copper Federal Water Quality Guideline.
}
\keyword{datasets}
