% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{All_WATER23_reactions}
\alias{All_WATER23_reactions}
\title{All WATER23.dbs reactions}
\format{
An object of class \code{list} of length 24.
}
\usage{
All_WATER23_reactions
}
\description{
A large problem using the WHAM V "WATER23.dbs" thermodynamic
  database. This is the thermodynamic database used in most of the Windows
  BLM parameter files, including the United States Environmental Protection
  Agency's (USEPA) final acute value (FAV).
}
\keyword{datasets}
