% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GetData.R
\name{BlankInputList}
\alias{BlankInputList}
\title{Make a blank inputs list object}
\usage{
BlankInputList(ThisProblem, NObs = 0L)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}
}
\value{
A list object with a template for defining the inputs for the
  chemical problem for the `BLMEngineInR` functions. Each element in the list
  is a vector, list, or data.frame object grouping related parameters
  together. See `str(BlankProblem())` for the structure and names of the list
  object.
}
\description{
This function is internal because the inputs can be specified in
  a data.frame object then matched to ThisProblem with
  `MatchInputsToProblem(DFInputs = , ThisProblem = )`. This function is
  useful as a comparison to make sure the outputs of `GetData` or
  `MatchInputsToProblem` and the inputs of `BLM` are acceptable.
}
\keyword{internal}
