% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankProblem.R
\name{BlankProblemList}
\alias{BlankProblemList}
\title{Make a blank input problem list object}
\usage{
BlankProblemList()
}
\value{
A list object where each element in the list is an input for BLM
  functions.
}
\description{
Make a blank input problem list object
}
\keyword{internal}
