% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BlankWHAM.R
\name{BlankWHAM}
\alias{BlankWHAM}
\title{Make a blank WHAM parameter list object}
\usage{
BlankWHAM()
}
\value{
A list object with a template for defining the organic matter binding
  in a chemical problem for the `BLMEngineInR` functions. Each element in the
  list is a vector, matrix, or data.frame object grouping related parameters
  together. See `str(BlankWHAM())` for the structure and names of the list
  object.
}
\description{
Make a blank WHAM parameter list object
}
