% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcIterationTotals}
\alias{CalcIterationTotals}
\title{Calculate the Moles and Concentrations For This Iteration}
\usage{
CalcIterationTotals(NComp, NSpec, SpecConc, SpecCtoM, SpecStoich, 
  CalcTotMoles, CalcTotConc);
}
\arguments{
\item{NComp}{integer, the number of components}

\item{NSpec}{integer, the number of species}

\item{SpecConc}{numeric vector (NSpec), the concentration of chemical
species}

\item{SpecStoich}{integer matrix (NSpec x NComp), the stoichiometry of
species formatin reactions}

\item{SpecCtoM}{numeric vector (NSpec), the concentration-to-mass conversion
factor for each chemical species}

\item{CalcTotConc}{(return parameter) numeric vector (NComp), the 
calculated total concentration of each component = CalcTotMoles / CtoM}

\item{CalcTotMoles}{(return parameter) numeric vector (NComp), the 
calculated total moles of each component = sum(SpecConc * SpecCtoM * 
SpecStoich[,j]) for each component j}
}
\value{
void
}
\description{
Sum the relevant species concentrations to calculate the 
  resulting total concentrations and moles for each component.
}
