% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcResidualAndError}
\alias{CalcResidualAndError}
\title{Calculate the Resid and CompError}
\usage{
CalcResidAndError(NComp, CalcTotMoles, TotMoles, CompType, Resid, CompError);
}
\arguments{
\item{NComp}{integer, the number of components}

\item{TotMoles}{numeric vector (NComp), the total moles of each component}

\item{CalcTotMoles}{numeric vector (NComp), the calculated total moles of 
each component = sum(SpecConc * SpecCtoM * SpecStoich[,j]) for each 
component j}

\item{CompType}{character vector (NComp), the type of component. It should
be a fixed set of values (MassBal, FixedAct, Substituted, ChargeBal,
SurfPot)}

\item{Resid}{(return value) numeric vector (NComp), the residuals = 
calculated totals - known totals}

\item{CompError}{(return value) numeric vector (NComp), the absolute error
  fraction for each component in this iteration = abs(Resid / TotMoles)

return void}
}
\description{
Calculate the residuals of the speciation problem, and the
  component Error.
}
\details{
The residuals of the speciation problem are the calculated total
  moles for each component minus their known concentrations. This function
  does not account for the different toxic metal residual needed in a 
  toxicity run.
}
