% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcResidualsOnly}
\alias{CalcResidualsOnly}
\title{Calculate the Residual - and only the residual}
\usage{
CalcResidualsOnly(NComp, CalcTotMoles, TotMoles, CompType)
}
\arguments{
\item{NComp}{integer, the number of components}

\item{TotMoles}{numeric vector (NComp), the total moles of each component}

\item{CalcTotMoles}{numeric vector (NComp), the calculated total moles of 
each component = sum(SpecConc * SpecCtoM * SpecStoich[,j]) for each 
component j}

\item{CompType}{character vector (NComp), the type of component. It should
be a fixed set of values (MassBal, FixedAct, Substituted, ChargeBal,
SurfPot)}
}
\value{
numeric vector (NComp), the residuals = calculated totals - known 
  totals
}
\description{
Calculate the residuals of the speciation problem.
}
\details{
The residuals of the speciation problem are the calculated total
  moles for each component minus their known concentrations. This function
  does not account for the different toxic metal residual needed in a 
  toxicity run.
}
