% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcStep}
\alias{CalcStep}
\title{Calculate the Newton-Raphson step}
\usage{
CalcStep(JacobianMatrix, Resid, NComp, CompType, CompName)
}
\arguments{
\item{JacobianMatrix}{numeric matrix (NComp x NComp), the Jacobian matrix
("Z")}

\item{Resid}{numeric vector (NComp), the residuals = calculated totals -
known totals}

\item{NComp}{integer, the number of components}

\item{CompType}{character vector (NComp), the type of component. It should be
a fixed set of values (MassBal, FixedAct, Substituted, ChargeBal, SurfPot)}

\item{CompName}{character vector (NComp), the names of the components}
}
\value{
numeric vector (NComp), the N-R step to take for each component ("X"
  in C(i+1) = C(i) - X)
}
\description{
Calculate the Newton-Raphson step
}
