% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{carbonate_system_problem}
\alias{carbonate_system_problem}
\title{Carbonate system problem}
\format{
An object of class \code{list} of length 22.
}
\usage{
carbonate_system_problem
}
\description{
An example BLMEngineInR problem object, which describes a
  water-only system with only the (closed) carbonate system.
}
\details{
This problem consists of two components (hydrogen "H" and carbonate
  "CO3") and three reactions (dissociation of water/formation of hydroxide
  "OH", formation of bicarbonate "HCO3" and formation of carbonic acid
  "H2CO3"). The pH and temperature are supplied as input variables, and the
  input label "ID" is supplied as well.
}
\examples{
print(carbonate_system_problem$Comp[, c("Name", "Charge", "Type")])
print(carbonate_system_problem$Spec[, c("Equation", "Charge", "ActCorr", "LogK", "DeltaH")])

}
\keyword{datasets}
