% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CHESS}
\alias{CHESS}
\title{CHemical Equilibria in Soils and Solutions}
\usage{
CHESS(
  QuietFlag,
  ConvergenceCriteria,
  MaxIter,
  NMass,
  MassName,
  MassAmt,
  NComp,
  CompName,
  CompType,
  TotConc,
  NSpec,
  SpecName,
  SpecType,
  SpecMCR,
  SpecK,
  SpecTempKelvin,
  SpecDeltaH,
  SpecStoich,
  SpecCharge,
  SpecActCorr,
  DoWHAM,
  AqueousMCR,
  WHAMDonnanMCR,
  HumicSubstGramsPerLiter,
  WHAMMolWt,
  WHAMRadius,
  WHAMP,
  WHAMDLF,
  WHAMKZED,
  SysTempKelvin,
  DoTox,
  MetalName,
  MetalCompR,
  BLCompR,
  NBLMetal,
  BLMetalSpecsR,
  CATarget,
  DodVidCj,
  DodVidCjDonnan,
  DodKidCj,
  DoGammai,
  DoJacDonnan,
  DoJacWHAM,
  DoWHAMSimpleAdjust,
  DoDonnanSimpleAdjust
)
}
\arguments{
\item{QuietFlag}{character, one of "Very Quiet" (only print out when run is
done), "Quiet" (print out Obs=iObs), or "Debug" (print out lots of info)}

\item{ConvergenceCriteria}{numeric, the maximum value of MaxError that 
counts as convergence by the Newton-Raphson root-finding algorithm}

\item{MaxIter}{integer, the maximum number of iterations the Newton-Raphson
root-finding algorithm should do before giving up}

\item{NMass}{integer, number of mass compartments}

\item{MassName}{CharacterVector (NMass), the names of the mass compartments}

\item{MassAmt}{NumericVector (NMass), The amount of each mass compartment.}

\item{NComp}{integer, number of components}

\item{CompName}{character vector (NComp), the name of each component in the
simulation}

\item{CompType}{character vector (NComp), the type of each component in the
simulation}

\item{TotConc}{numeric vector (NComp), the total concentrations of each
component in the simulation (units of e.g., mol/L and mol/kg)}

\item{NSpec}{integer, number of species reactions}

\item{SpecName}{character vector (NSpec), the name of the chemical species
for which we have formation reactions}

\item{SpecType}{character vector (NSpec), the type or category of the 
chemical species for which we have formation reactions.}

\item{SpecMCR}{IntegerVector (NSpec), the mass compartment of the chemical
species for which we have formation reactions}

\item{SpecK}{numeric vector (NSpec), the equilibrium coefficient of the
formation reactions.}

\item{SpecTempKelvin}{NumericVector (NSpec), the temperature associated with
K/logK and DeltaH of the formation reactions}

\item{SpecDeltaH}{numeric vector (NSpec), the enthalpy change of the
formation reactions}

\item{SpecStoich}{signed integer matrix (NSpec x NComp), the reaction
stoichiometry of the formation reactions}

\item{SpecCharge}{signed integer vector (NSpec), the charge of the chemical
species for which we have formation reactions}

\item{SpecActCorr}{character vector (NSpec), the activity correction method
of the chemical species for which we have formation reactions}

\item{DoWHAM}{boolean, true=there are WHAM species, false=no WHAM species}

\item{AqueousMCR}{integer, the (1-based) position of the water/aqueous mass
compartment. (transformed to 0-based at the beginning of the function)}

\item{WHAMDonnanMCR}{the mass compartments corresponding to the
humic acid (0) and fulvic acid (1) Donnan layers. (transformed to 0-based
at the beginning of the function)}

\item{HumicSubstGramsPerLiter}{NumericVector, length of 2, grams per liter 
of each organic matter component (HA and FA) in solution}

\item{WHAMMolWt}{numeric (2), WHAM's molecular weight parameter for organic
matter}

\item{WHAMRadius}{numeric (2), WHAM's molecular radius parameter for organic
matter}

\item{WHAMP}{numeric (2), WHAM's P parameter...}

\item{WHAMDLF}{numeric (2), WHAM's Double layer overlap factor}

\item{WHAMKZED}{numeric (2), WHAM's Constant to control DDL at low ZED}

\item{SysTempKelvin}{double; input temperature for the current observation,
in Kelvin}

\item{DoTox}{logical, TRUE for toxicity mode where the MetalName component
concentration is adjusted to try to match the CATarget with BLMetalSpecs}

\item{MetalName}{character string, the name of the toxic metal}

\item{MetalCompR}{integer, the position of the metal in the component arrays
(i.e., which is the toxic metal component) Note: this is base-1 indexed on
input then converted.}

\item{NBLMetal}{integer, the number of biotic ligand-bound metal species 
that are associated with toxic effects.}

\item{BLMetalSpecsR}{integer vector, the positions of the species in the
arrays which contribute to toxicity (i.e., which species are the toxic
metal bound to the relevant biotic ligand) Note: these are base-1 indexed
on input then converted.}

\item{CATarget}{numeric, the target critical accumulation in units of mol /
kg (only used when DoTox == TRUE)}
}
\value{
list with the following elements:
\describe{
  \item{SpecConc}{numeric vector (NSpec), the concentrations of each species
    for which we have formation reactions}
  \item{FinalIter}{integer, the number of Newton-Raphson iterations that we
    needed to reach convergence}
  \item{FinalMaxError}{numeric, the highest final absolute error fraction  
    =max(abs(Resid / TotMoles))}
  \item{CalcTotConc}{numeric vector (NComp), the calculated total
    concentrations of each component in the simulation (units of e.g., mol/L
    and mol/kg)}
}
}
\description{
Given a chemical system, equilibria equations, and total
  concentrations of components, calculate the species concentrations of each
  chemical product in the system.
}
