% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CompUpdate}
\alias{CompUpdate}
\title{Iterative step improvement in component concentrations}
\usage{
CompUpdate(NComp, CompConcStep, CompConc, CompName)
}
\arguments{
\item{NComp}{integer, the number of components}

\item{CompConcStep}{numeric vector (NComp) of adjustments to the component
concentrations}

\item{CompConc}{numeric vector (NComp) of component concentrations, input values
are from this iteration}

\item{CompName}{numeric vector (NComp) with the names of the components}
}
\value{
numeric cector CompConc (NComp) modified for the next iteration
}
\description{
CompUpdate calculates an iterative improvement on the component
concentrations based on the Newton-Raphson solution from the current
iteration.
}
\details{
If the iteration would cause the adjusted component concentrations
to be less than zero, then the component concentration is simply divided by
10 for this iteration.
}
