% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Cu_full_organic_problem}
\alias{Cu_full_organic_problem}
\title{Copper problem with WHAM V organic matter}
\format{
An object of class \code{list} of length 23.
}
\usage{
Cu_full_organic_problem
}
\description{
An example BLMEngineInR problem object, which describes a system
  with organic matter represented by WHAM V, and all of the common cations
  (Ca, Mg, Na, K) and anions (SO4, Cl, CO3) represented with their usual
  reactions. Copper is also represented as the toxic metal binding to a
  biotic ligand, and some example critical accumulations values are provided,
  including one for the United States Environmental Protection Agency's
  (USEPA) final acute value (FAV).
}
\keyword{datasets}
