% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DefineProblem.R
\name{DefineProblem}
\alias{DefineProblem}
\title{Define the speciation problem}
\usage{
DefineProblem(ParamFile)
}
\arguments{
\item{ParamFile}{the path and file name to a parameter file}
}
\value{
Returns a `list` object with everything that is needed to
  characterize the equilibrium system being solved. Each list item named
  according to the Common Parameter Definitions used throughout the package.
}
\description{
`DefineProblem` reads in a parameter file, and sets up the required vectors
and matrices that will be needed to run the speciation calculations in CHESS.
}
\examples{
## Not Run
# thisProblem = DefineProblem("my_parameter_file.dat")
## End Not Run
}
\seealso{
Other BLMEngine Functions: 
\code{\link{CommonParameterDefinitions}},
\code{\link{ReadInputsFromFile}()}
}
\concept{BLMEngine Functions}
