% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Full_InorgDataTotalConc}
\alias{Full_InorgDataTotalConc}
\alias{Full_InorgDataFreeConc}
\alias{Full_InorgDataK}
\alias{Full_InorgDataStoich}
\alias{Full_InorgDataCharge}
\title{Simple Test DAta with Full Inorganic Components}
\format{
\code{Full_InorgDataTotalConc} is an object of class \code{numeric} of
  length 11 containing total concentrations.

\code{Full_InorgDataFreeConc} is an object of class \code{numeric} of
  length 32 containing free concentrations.

\code{Full_InorgDataK} is an object of class \code{numeric} of
  length 31 containing equilibrium coefficients.

\code{Full_InorgDataStoich} is An object of class \code{matrix} (inherits
  from \code{array}) with 32 rows and 11 columns containing stoichiometry
  information for each reaction.
  \describe{
    \item{\code{rows}}{Species reactions.}
    \item{\code{cols}}{Components.}
  }

\code{Full_InorgDataCharge} is an object of class \code{integer} of
  length 32 containing species charges.
}
\source{
Windward BLM (Ver 3.57.2.49, build 2022-03-30)
}
\usage{
Full_InorgDataTotalConc

Full_InorgDataFreeConc

Full_InorgDataK

Full_InorgDataStoich

Full_InorgDataCharge
}
\description{
A collection of chemistry inputs and outputs from an existing
  BLM run with all inorganic components (H, Cu, DOC, Ca, Mg, Na, K, SO4, Cl,
  CO3, BL).
}
\details{
Windward BLM (Ver 3.57.2.49, build 2022-03-30)\cr Parameter File:
  Cu_freshwater_acute_and_chronic_2017-01-17.dat\cr Switches: /W /Q /VER3.57
  /O3\cr
\itemize{
   \item Temp. = 25
   \item HA = 10
   \item pH = 7
   \item Cu = 1.57366E-028
   \item DOC = 1E-008
   \item Ca = 2.49513E-025
   \item Mg = 4.11438E-025
   \item Na = 4.34976E-025
   \item K = 2.55766E-025
   \item SO4 = 1.041E-025
   \item Cl = 2.82064E-025
   \item DIC = 0.0001
   \item S = 3.11876E-025
 }
}
\keyword{datasets}
