% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Full_OrganicDataTotalConc}
\alias{Full_OrganicDataTotalConc}
\alias{Full_OrganicDataFreeConc}
\title{Simple Test Data with Full Organic Components}
\format{
\code{Full_OrganicDataTotalConc} is an object of class \code{numeric} of
  length 11 containing total concentrations.

\code{Full_OrganicDataFreeConc} is an object of class \code{numeric} of
  length 38 containing free concentrations.
}
\source{
Windward BLM (Ver 3.57.2.49, build 2022-03-30)
}
\usage{
Full_OrganicDataTotalConc

Full_OrganicDataFreeConc
}
\description{
A collection of chemistry inputs and outputs from an existing
  BLM run with all organic components (H, Cu, DOC, Ca, Mg, Na, K, SO4, Cl,
  CO3, BL).
}
\details{
Windward BLM (Ver 3.57.2.49, build 2022-03-30)\cr Parameter File:
  Cu_freshwater_acute_and_chronic_2017-01-17.dat\cr Switches: /W /Q /VER3.57
  /O3\cr
\itemize{
   \item Temp = 20 deg C
   \item pH = 7.6 SU
   \item Cu = 3 ug/L
   \item DOC = 2 mg C/L
   \item HA = 0.001	%
   \item Ca = 12	mg/L
   \item Mg = 12.2	mg/L
   \item Na = 25.3	mg/L
   \item K = 2.1	mg/L
   \item SO4 = 76.8 mg/L
   \item Cl = 1.9 mg/L
   \item Alk = 52.2 mg/L as CaCO3
   \item S = 1E-009 mg/L
 }
}
\keyword{datasets}
