% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Jacobian}
\alias{Jacobian}
\title{Calculate the Jacobian matrix}
\usage{
Jacobian(
  NComp,
  NSpec,
  SpecStoich,
  SpecConc,
  SpecCtoM,
  CompName,
  MetalComp,
  NBLMetal,
  BLMetalSpecs,
  DoTox
)
}
\arguments{
\item{NComp}{integer, the number of components}

\item{NSpec}{integer, the number of species}

\item{SpecStoich}{integer matrix (NSpec x NComp), the stoichiometry of
each reaction}

\item{SpecConc}{numeric vector (NSpec), the concentrations of each species}

\item{SpecCtoM}{numeric vector (NSpec), the factor to apply to
concentrations (i.e., units of mol/L or mol/kg) to convert to masses (i.e.,
units of mol).}

\item{MetalComp}{integer, the position in component vectors of the toxic
metal component}

\item{BLMetalSpecs}{integer vector, the position in the species vectors of
the metal-biotic ligand species associated with toxicity}

\item{DoTox}{logical, if TRUE = toxicity run, FALSE = speciation run}

\item{SpecName}{character vector (NSpec), the species names}
}
\value{
numeric matrix (NComp x NComp), the jacobian matrix (see details)
}
\description{
Calculate the Jacobian matrix of the speciation problem
}
\details{
This function calculates the Jacobian matrix of the speciation
  problem. The Jacobian matrix is an (`NComp` x `NComp`) matrix of
  derivatives, where each row gives the derivative of the residual each total
  component concentration, dR[i], with respect to, in each column, the
  component free concentration, dX[j]. (i.e., \eqn{Z_{i,j} = \delta Resid_{i}
  / \delta CompConc_{j}}).
}
\keyword{internal}
