% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listCAT.R
\name{listCAT}
\alias{listCAT}
\title{List Critical Accumulation Table}
\usage{
listCAT(paramFile)
}
\arguments{
\item{paramFile}{the file name and path of the parameter file.}
}
\value{
A data frame with species and test information, and the CAT numbers
  that would correspond to each
}
\description{
List out the critical accumulation table for the user to allow them to pick
which CAT number they should specify for a toxicity run where the critical
value is coming from the table in the parameter file.
}
\examples{
## Not Run
# listCAT
## End Not Run
}
