% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{MaxCompError}
\alias{MaxCompError}
\title{Find the MaxError and WhichMax}
\usage{
MaxError = MaxCompError(NComp, CompError, WhichMax)
}
\arguments{
\item{NComp}{integer, the number of components}

\item{CompError}{numeric vector (NComp), the absolute error fraction for
each component in this iteration =abs(Resid / TotMoles)}

\item{MaxError}{(return value) numeric, the highest absolute error fraction in this
iteration =max(abs(Resid / TotMoles))}

\item{WhichMax}{(return value) integer, the position in the component vectors of the
component with the highest absolute error}
}
\value{
void
}
\description{
Determine which component has the maximum absolute error 
  fraction.
}
