% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Ni_HCO3_full_organic_problem}
\alias{Ni_HCO3_full_organic_problem}
\title{Ni problem with WHAM V organic matter and NiHCO3 toxic}
\format{
An object of class \code{list} of length 25.
}
\usage{
Ni_HCO3_full_organic_problem
}
\description{
An example BLMEngineInR problem object, which describes a system
  with all of the common cations (Ca, Mg, Na, K) and anions (SO4, Cl, CO3)
  represented with their usual reactions. Nickel is also represented as the
  toxic metal binding to the biotic ligand, and the critical accumulations
  from the Santore et al. (2021) paper. This version has a BL-NiHCO3 species
  whose binding constant has been calibrated to Ceriodaphnia dubia toxicity.
  Ceriodaphnia dubia are sensitive to bicarbonate toxicity and this file
  simulates this mixtures effect.
}
\references{
Santore, Robert C., Kelly Croteau, Adam C. Ryan, Christian Schlekat,
Elizabeth Middleton, Emily Garmin, and Tham Hoang (2021). A Review of Water
Quality Factors that Affect Nickel Bioavailability to Aquatic Organisms:
Refinement of the Biotic Ligand Model for Nickel in Acute and Chronic
Exposures. Environmental Toxicology and Chemistry, col 40, iss. 8, pp
2121-2134. doi: 10.1002/etc.5109
}
\keyword{datasets}
