% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemConversionFunctions.R
\name{ProblemConversionFunctions}
\alias{ProblemConversionFunctions}
\alias{ConvertToList}
\alias{ConvertToDF}
\title{Problem Conversion functions}
\usage{
ConvertToList(ThisProblemDF)

ConvertToDF(ThisProblemList)
}
\arguments{
\item{ThisProblemDF}{a list object with data.frames, lists, and vectors,
following the template given by `BlankProblem()`.}

\item{ThisProblemList}{a list object where each element is an input for
CHESS, following the template given by `BlankProblemList()`.}
}
\value{
a list object of the opposite type
}
\description{
These functions are for converting between a data.frame-based
  list object and a pure list-based object. The data.frame-based list object
  is easier to edit, generally, while the pure list-based object is more
  readily used by C++ functions such as CHESS.
}
\keyword{internal}
