% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCalcResidual.R
\name{RCalcResidual}
\alias{RCalcResidual}
\title{Calculate the Residual}
\usage{
RCalcResidual(
  NComp,
  NSpec,
  SpecConc,
  SpecStoich,
  TotMoles,
  SpecCtoM,
  CompName,
  CompType,
  MetalComp,
  BLMetalSpecs,
  CATarget,
  DoTox
)
}
\arguments{
\item{DoTox}{}
}
\description{
Calculate the residuals of the speciation problem.
}
\details{
The residuals of the speciation problem are the calculated total concentrations \eqn{\Sum()}
}
