% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCalcSpecConc.R
\name{RCalcSpecConc}
\alias{RCalcSpecConc}
\title{Calculate Species Concentrations}
\usage{
RCalcSpecConc(
  CompConc,
  SpecK,
  SpecStoich,
  NComp = length(CompConc),
  NSpec = length(SpecK),
  SpecName = names(SpecK)
)
}
\arguments{
\item{CompConc}{A vector of component concentrations for each of `NComp`
components.}

\item{SpecK}{A vector of reaction equilibrium constants for each of `NSpec`
reactions.}

\item{SpecStoich}{A matrix of reaction stoichiometry, with `NSpec` rows and
`NComp` columns.}

\item{NComp}{The number of components in the equilibrium system.}

\item{NSpec}{The number of species (reactions) in the equilibrium system.}

\item{SpecName}{character vector (NSpec), the name of the chemical species
for which we have formation reactions}
}
\value{
A vector of `NSpec` species concentrations.
}
\description{
Use `RCalcSpecConc` to calculate species concentrations from a known set of
free component ion concentrations.
}
\details{
This is an internal function that will, for each species `i` in `NSpec`
calculate the equilibrium concentration, which in its most basic form would
be calculated as \eqn{SpecConc_{i} = K_{i} *
\prod_{j=1}^{n}(CompConc^SpecStoich_{i,j})}. Further modifications to these
calculations would include temperature correction, ionic strength
corrections, and diffuse double layer calculations for organic matter
binding. As of 19-Oct-2023, none of these corrections have been implemented.

This is the base-R version of this function.
}
