% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCHESS.R
\name{RCHESS}
\alias{RCHESS}
\title{CHemical Equilibria in Soils and Solutions}
\usage{
RCHESS(
  DoPartialSteps,
  QuietFlag,
  ConvergenceCriteria,
  MaxIter,
  NComp,
  NSpec,
  NBLMetal,
  SpecK,
  SpecStoich,
  SpecCtoM,
  SpecName,
  CompType,
  CompName,
  TotMoles,
  TotConc,
  DoTox,
  MetalName,
  MetalComp,
  BLMetalSpecs,
  CATarget
)
}
\arguments{
\item{DoPartialSteps}{logical, TRUE = do partial Newton-Raphson steps when a
full step would send the MaxError higher; FALSE = only do full steps}

\item{QuietFlag}{character, one of "Very Quiet" (only print Out when run is
done), "Quiet" (print Out Obs=iObs), or "Debug" (print Out lots of info)}

\item{ConvergenceCriteria}{numeric, the maximum value of MaxError that counts
as convergence by the Newton-Raphson root-finding algorithm}

\item{MaxIter}{integer, the maximum number of iterations the Newton-Raphson
root-finding algorithm should do before giving up}

\item{NComp}{integer, number of components}

\item{NSpec}{integer, number of species reactions}

\item{NBLMetal}{integer, the number of biotic ligand-bound metal species that
are associated with toxic effects.}

\item{SpecK}{numeric vector (NSpec), the equilibrium coefficient of the
formation reactions}

\item{SpecStoich}{signed integer matrix (NSpec x NComp), the reaction
stoichiometry of the formation reactions}

\item{SpecCtoM}{numeric vector (NSpec), the concentration to mass conversion
factor of the chemical species for which we have formation reactions}

\item{SpecName}{character vector (NSpec), the name of the chemical species
for which we have formation reactions}

\item{CompType}{character vector (NComp), the type of each component in the
simulation}

\item{CompName}{character vector (NComp), the name of each component in the
simulation}

\item{TotMoles}{numeric vector (NComp), the total moles of each component in
the simulation (units of mol)}

\item{TotConc}{numeric vector (NComp), the total concentrations of each
component in the simulation (units of e.g., mol/L and mol/kg)}

\item{DoTox}{logical, TRUE for toxicity mode where the MetalName component
concentration is adjusted to try to match the CATarget with BLMetalSpecs}

\item{MetalName}{character string, the name of the toxic metal}

\item{MetalComp}{integer, the position of the metal in the component arrays
(i.e., which is the toxic metal component) Note: this are base-1 indexed.}

\item{BLMetalSpecs}{integer vector, the positions of the species in the
arrays which contribute to toxicity (i.e., which species are the toxic
metal bound to the relevant biotic ligand) Note: these are base-1 indexed.}

\item{CATarget}{numeric, the target critical accumulation in units of mol /
kg (only used when DoTox == TRUE)}
}
\value{
list with the following elements:
  \desribe{
    \item{SpecConc}{numeric vector (NSpec), the concentrations of each
      species for which we have formation reactions}
    \item{Iter}{integer, the number of Newton-Raphson iterations that we
      needed to reach convergence}
    \item{MaxError}{numeric, the highest final absolute error fraction
      =max(abs(Resid / TotMoles))}
    \item{CalcTotConc}{numeric vector (NComp), the calculated total
      concentrations of each component in the simulation (units of e.g.,
      mol/L and mol/kg)}
  }
}
\description{
Given a chemical system, equilibria equations, and total
  concentrations of components, calculate the species concentrations of each
  chemical product in the system.
}
