% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCompUpdate.R
\name{RCompUpdate}
\alias{RCompUpdate}
\title{Iterative step improvement in component concentrations}
\usage{
RCompUpdate(CompConcStep, CompConc, CompName)
}
\arguments{
\item{CompConcStep}{: Vector (NComp) of adjustments to the component
concentrations}

\item{CompConc}{: Vector (NComp) of component concentrations, input values
are from this iteration}

\item{CompName}{: Vector (NComp) with the names of the components}
}
\value{
Vector CompConc (NComp) modified for the next iteration
}
\description{
RCompUpdate calculates an iterative improvement on the component
concentrations based on the Newton-Raphson solution from the current
iteration.
}
\details{
If the iteration would cause the adjusted component concentrations
to be less than zero, then the component concentration is simply divided by
10 for this iteration.
}
