% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{RemoveComponents}
\alias{RemoveComponents}
\title{Remove a component from a problem list object}
\usage{
RemoveComponents(ThisProblem, ComponentToRemove)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{ComponentToRemove}{A character vector with names or indices of the
component(s) to remove from `ThisProblem`.}
}
\value{
`ThisProblem`, minus the component(s) indicated by
  `ComponentToRemove`, also with any species, etc associates with those
  components removed and with any other ripple effects updated.
}
\description{
Remove a component from a problem list object
}
