% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{RemoveDefComps}
\alias{RemoveDefComps}
\title{Remove a defined component from a problem list object}
\usage{
RemoveDefComps(ThisProblem, DefCompToRemove)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{DefCompToRemove}{A character vector with names or indices of the
defined component(s) to remove from `ThisProblem`.}
}
\value{
`ThisProblem`, minus the defined component(s) indicated by
  `DefCompToRemove`, also with any other ripple effects updated.
}
\description{
Remove a defined component from a problem list object
}
