% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{RemoveInLab}
\alias{RemoveInLab}
\title{Remove an input label from a problem list object}
\usage{
RemoveInLab(ThisProblem, InLabToRemove)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{InLabToRemove}{A character vector with names or indices of the input
label(s) to remove from `ThisProblem`.}
}
\value{
`ThisProblem`, minus the input label(s) indicated by
  `InLabToRemove`.
}
\description{
Remove an input label from a problem list object
}
