% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{RemoveInVar}
\alias{RemoveInVar}
\title{Remove an input variable from a problem list object}
\usage{
RemoveInVar(ThisProblem, InVarToRemove)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{InVarToRemove}{A character vector with names or indices of the input
variable(s) to remove from `ThisProblem`.}
}
\value{
`ThisProblem`, minus the input variable(s) indicated by
  `InVarToRemove`, also with any components, etc. associated with those
  variables.
}
\description{
Remove an input variable from a problem list object
}
