% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{RemoveMassCompartments}
\alias{RemoveMassCompartments}
\title{Remove a Mass Compartment from a problem list object}
\usage{
RemoveMassCompartments(ThisProblem, MCToRemove)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{MCToRemove}{A character vector with names or indices of the mass
compartment(s) to remove from `ThisProblem`.}
}
\value{
`ThisProblem`, minus the mass compartment(s) indicated by
  `MCToRemove`, also with any components, input variables, etc. associated
  with those mass compartments removed.
}
\description{
Remove a Mass Compartment from a problem list object
}
