% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ProblemManipulation.R
\name{RemoveSpecies}
\alias{RemoveSpecies}
\title{Remove a species reaction from a problem list object}
\usage{
RemoveSpecies(ThisProblem, SpeciesToRemove)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{SpeciesToRemove}{A character vector with names or indices of the
species(s) to remove from `ThisProblem`.}
}
\value{
`ThisProblem`, minus the species indicated by
  `SpeciesToRemove`, also with any other ripple effects updated.
}
\description{
Remove a species reaction from a problem list object
}
