% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TestDataK}
\alias{TestDataK}
\title{Reaction Equilibrium Constants for Test Data}
\format{
An object of class \code{numeric} of length 4.
}
\source{
Windward BLM (Ver 3.57.2.49, build 2022-03-30)
}
\usage{
TestDataK
}
\description{
Reaction equilibrium constants for species reactions in a
  simple carbonate equilibrium system. The only components are hydrogen (H)
  and carbonate (CO3). The only species are H(+), CO3(-2), HCO3(-), and
  H2CO3(0).
}
\details{
Windward BLM (Ver 3.57.2.49, build 2022-03-30)\cr
Parameter File: Cu_freshwater_acute_and_chronic_2017-01-17.dat\cr
Switches: /W /Q /VER3.57 /O3\cr
\itemize{
   \item Temp. = 25
   \item HA = 10
   \item pH = 7
   \item Cu = 1.57366E-028
   \item DOC = 1E-008
   \item Ca = 2.49513E-025
   \item Mg = 4.11438E-025
   \item Na = 4.34976E-025
   \item K = 2.55766E-025
   \item SO4 = 1.041E-025
   \item Cl = 2.82064E-025
   \item DIC = 0.0001
   \item S = 3.11876E-025
 }
}
\keyword{datasets}
