% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{UpdateTotalsList}
\alias{UpdateTotalsList}
\title{Update Fixed Activity and Metal Totals}
\usage{
UpdateTotalsList(
  NComp,
  NSpec,
  DoTox,
  CompType,
  CompName,
  MetalName,
  SpecStoich,
  SpecMoles,
  CompCtoM,
  TotMoles,
  TotConc
)
}
\arguments{
\item{NComp}{integer, the combined number of components in the simulation, 
including the input components, defined components (and including the 
defined components that get added by ExpandWHAM)}

\item{NSpec}{integer, the number of chemical species for which we have 
formation reactions in the simulation}

\item{DoTox}{logical, TRUE for toxicity mode where the MetalName component 
concentration is adjusted to try to match the CATarget with BLMetalSpecs}

\item{CompType}{character vector (NComp), the type of each component in the 
simulation}

\item{CompName}{character vector (NComp), the name of each component in the 
simulation}

\item{MetalName}{The name of the component that corresponds to the metal 
associated with toxic effects.}

\item{SpecStoich}{signed integer matrix (NSpec x NComp), the reaction 
stoichiometry of the formation reactions.}

\item{SpecMoles}{numeric vector (NSpec), the moles of each species for 
which we have formation reactions}

\item{CompCtoM}{numeric vector (NSpec), the concentration to mass conversion 
factor of the components}

\item{TotMoles}{numeric vector (NComp), the total moles of each component in 
the simulation (units of mol)}

\item{TotConc}{numeric vector (NComp), the total concentrations of each 
component in the simulation (units of e.g., mol/L and mol/kg)}
}
\value{
Rcpp::List (Rcpp::NumericVector TotMoles, 
  Rcpp::NumericVector TotConc)
}
\description{
Update the totals that change with each iteration - the totals 
  for fixed activity components, and the metal component during tox mode. 
  (returns a list)
}
