% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_MC_problem}
\alias{water_MC_problem}
\title{Water mass compartment only problem}
\format{
An object of class \code{list} of length 24.
}
\usage{
water_MC_problem
}
\description{
An example BLMEngineInR problem object, which describes a
  water-only system with no input variables or components yet, and the input
  label "ID".
}
\keyword{datasets}
