% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{water_problem}
\alias{water_problem}
\title{Water-only problem}
\format{
An object of class \code{list} of length 23.
}
\usage{
water_problem
}
\description{
An example BLMEngineInR problem object, which describes a
  water-only system with pH and temperature are supplied as input variables,
  and the input label "ID" is supplied as well. The only reaction is water
  dissociation (hydroxide "OH" formation reaction).
}
\keyword{datasets}
