% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteDetailedFile.R
\name{WriteDetailedFile}
\alias{WriteDetailedFile}
\title{Write a VERY Detailed Output File}
\usage{
WriteDetailedFile(
  OutList,
  FileName,
  AdditionalInfo = paste0("Saved on: ", Sys.time())
)
}
\arguments{
\item{OutList}{The list object returned by the BLM function.}

\item{FileName}{The name of the file you'd like to write.}

\item{AdditionalInfo}{This vector will be included in the "Additional Info".
By default, it will give the date/time the file was saved.}
}
\value{
Returns TRUE (invisibly) if successful.
}
\description{
This will write an output XLSX file with everything that is returned by the
`BLM` function. This includes inputs, concentrations, activities, etc.
}
