% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteParamFile.R
\name{WriteParamFile}
\alias{WriteParamFile}
\title{Write a BLM Parameter File}
\usage{
WriteParamFile(ThisProblem, ParamFile)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{ParamFile}{a character value, indicating the file path and name of the
parameter file to write.}
}
\value{
(invisibly) TRUE, if successful.
}
\description{
This function will take a BLM chemical problem list object and
  turn it into a parameter file, effectively doing the opposite of
  `DefineProblem`.
}
\examples{
## Not run:

myfile_problem = BLMEngineInR::DefineProblem(ParamFile = "myfile.dat4")

BLMEngineInR::WriteParamFile(ThisProblem = myfile_problem, ParamFile = "myfile_copy.dat4")

# myfile.dat4 and myfile_copy.dat4 should be nearly identical, except for white space

## End (Not run )

}
