% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLM.R
\name{BLM}
\alias{BLM}
\title{Run the Biotic Ligand Model}
\usage{
BLM(paramFile, inputFile)
}
\arguments{
\item{paramFile}{the path and file name of the parameter file}

\item{inputFile}{the path and file name of the chemistry input file}
}
\value{
A data frame with chemistry speciation information, including total
  concentrations.
}
\description{
`BLM` will run the Windward Environmental Biotic Ligand Model (BLM) with the
provided parameter file, input file, and options.
}
\examples{
## Not run:
# BLM(paramFile = "path/mypfile.dat", inputFile = "path/myinputfile.blm")
## End(Not run)
}
