% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/BLMEngineInR-package.R
\docType{package}
\name{BLMEngineInR-package}
\alias{BLMEngineInR}
\alias{BLMEngineInR-package}
\title{BLMEngineInR: BLM Engine In R}
\description{
The Windward Environmental, LLC Biotic Ligand Model, now as an R package. Originally programmed in PowerBasic by Robert Santore and others. WORK IN PROGRESS.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://www.windwardenv.com/biotic-ligand-model/}
}

}
\author{
\strong{Maintainer}: Kelly Croteau \email{kellyc@windwardenv.com}

Authors:
\itemize{
  \item Robert Santore \email{roberts@windwardenv.com}
}

}
\keyword{internal}
