% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/defineProblem.R
\name{defineProblem}
\alias{defineProblem}
\title{Define the speciation problem}
\usage{
defineProblem(paramFile)
}
\arguments{
\item{paramFile}{the path and file name to a parameter file}
}
\value{
Returns whatever is needed for the speciation problem to be run. A
`list` object with the following components:
\describe{
 \item{\code{NComp}}{integer; the number of components}
 \item{\code{NSpec}}{integer; the number of species}
 \item{\code{CompName}}{character vector of length `NComp`; component names}
 \item{\code{CompCharge}}{integer vector of length `NComp`; the charge of the components as free ions}
 \item{\code{CompType}}{character vector of length `NComp`; the type of component. It should be a fixed set of values (MassBal, FixedAct, Substituted, ChargeBal, SurfPot)}
 \item{\code{CompActCorr}}{character vector of length `NComp`; the method to use for activity corrections with this component,  }
 \item{\code{SpecName}}{character vector of length `NSpec`; species names}
 \item{\code{SpecActCorr}}{integer vector of length `NSpec`; the method to use for activity corrections with this speies where 1 = ...}
 \item{\code{Stoich}}{the stoichiometry matrix of the reactions (integer matrix of size `NSpec` x `nComp`)}
 \item{\code{K}}{the equilibrium coefficients (numeric vector of length `NSpec`)}
 \item{\code{logK}}{the log10-transformed equilibrium coefficients (numeric vector of length `NSpec`)}
 \item{\code{DeltaH}}{the enthalpy change for each formation reaction (species)}
}
}
\description{
`defineProblem` reads in a parameter file, and sets up the required vectors
and matrices that will be needed to run the speciation calculations in CHESS.
}
\examples{
## Not Run
# thisProblem = defineProblem("my_parameter_file.dat")
## End Not Run
}
