% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{GetData}
\alias{GetData}
\title{Get data from the input file}
\usage{
GetData(
  InputFile,
  NInLab,
  InLabName,
  NInVar,
  InVarName,
  InVarMCR,
  InVarType,
  NInComp,
  InCompName,
  NComp,
  CompName,
  NDefComp,
  DefCompName,
  DefCompFromNum,
  DefCompFromVar,
  DefCompSiteDens
)
}
\arguments{
\item{InputFile}{character(1); the path and file name to a BLM input file}

\item{NInLab}{integer; number of input label columns}

\item{InLabName}{character vector of length `NInLab`; names of input columns}

\item{NInVar}{integer; Number of input variables}

\item{InVarName}{character vector of length `NInVar`; Names of input
variables}

\item{InVarMCR}{integer vector of length `NInVar`;  Mass compartments of input
variables}

\item{InVarType}{character vector of length `NInVar`; Types of input
variables}

\item{NInComp}{integer; Number in input components}

\item{InCompName}{character vector of length `NInComp`; names of input
components}

\item{NComp}{integer; Number of components}

\item{CompName}{character vector of length `NComp`; component names}

\item{NDefComp}{integer; Number of defined components}

\item{DefCompName}{character vector of length `NDefComp`; defined component
names}

\item{DefCompFromNum}{numeric vector of length `NDefComp`; the number the
defined component is formed from}

\item{DefCompFromVar}{character vector of length `NDefComp`; the column used
to form the defined component}

\item{DefCompSiteDens}{numeric vector of length `NDefComp`; the binding site
density of each defined component}
}
\value{
Returns a `list` object with the following components:
\describe{
  \item{NObs}{integer; the number of chemistry observations}
  \item{InLabObs}{matrix with NObs rows and InLab columns; the input labels
    for each observation}
  \item{InVarObs}{matrix with NObs rows and InVar columns; the input
    variables for each observation}
  \item{InCompObs}{matrix with NObs rows and InComp columns; the input
    component concentrations for each observation}
  \item{SysTempCelsiusObs}{numeric vector of length NObs; input temperatures,
    in Celsius}
  \item{pH}{numeric vector NObs; input pH for each observation}
  \item{TotConcObs}{numeric matrix with NObs rows and NComp columns; the
    total concentrations of each component, including derived components}
}
}
\description{
`GetData` reads in the input file and prepares it for input to the BLM function.
}
\seealso{
Other BLMEngine Functions: 
\code{\link{CommonParameterDefinitions}},
\code{\link{MatchInputsToProblem}()},
\code{\link{ReadInputsFromFile}()}
}
\concept{BLMEngine Functions}
