% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RCalcResidual.R
\name{RCalcResidual}
\alias{RCalcResidual}
\title{Calculate the Residual}
\usage{
RCalcResidual(
  NComp,
  NSpec,
  SpecConc,
  SpecStoich,
  TotMoles,
  SpecCtoM,
  CompName,
  CompType,
  MetalComp,
  BLMetalSpecs,
  CATarget,
  DoTox
)
}
\arguments{
\item{NComp}{integer, the number of components}

\item{NSpec}{integer, the number of species}

\item{SpecConc}{numeric vector (NSpec), the concentration of chemical species}

\item{SpecStoich}{integer matrix (NSpec x NComp), the stoichiometry of
species formatin reactions}

\item{TotMoles}{numeric vector (NComp), the total moles of each component}

\item{SpecCtoM}{numeric vector (NSpec), the concentration-to-mass conversion
factor for each chemical species}

\item{CompName}{character vector (NComp), the names of the components}

\item{CompType}{character vector (NComp), the type of component. It should be
a fixed set of values (MassBal, FixedAct, Substituted, ChargeBal, SurfPot)}

\item{MetalComp}{integer, the position in component vectors of the toxic
metal component}

\item{BLMetalSpecs}{integer vector, the position in the species vectors of
the metal-biotic ligand species associated with toxicity}

\item{CATarget}{numeric, the target critical accumulation value for a
toxicity run, in mol/kg}

\item{DoTox}{logical, if TRUE = toxicity run, FALSE = speciation run}
}
\value{
A `list` object with the following components:
\describe{
 \item{\code{MaxError}}{numeric, the highest absolute error fraction in this
  iteration =max(abs(Resid / TotMoles))}
 \item{\code{WhichMax}}{integer, the position in the component vectors of the
   component with the highest absolute error}
 \item{\code{Resid}}{numeric vector (NComp), the residuals =
   calculated totals - known totals}
 \item{\code{CompError}}{numeric vector (NComp), the absolute error fraction
   for each component in this iteration =abs(Resid / TotMoles)}
 \item{\code{CalcTotConc}}{numeric vector (NComp), the calculated total
   concentration of each component = CalcTotMoles / CtoM}
 \item{\code{CalcTotMoles}}{numeric vector (NComp), the calculated total
   moles of each component = sum(SpecConc * SpecCtoM * SpecStoich[,j]) for
   each component j}
}
}
\description{
Calculate the residuals of the speciation problem.
}
\details{
The residuals of the speciation problem are the calculated total
  moles for each component minus their known concentrations. In the case of a
  toxicity run (DoTox = TRUE), the residual for the `MetalComp` component is
  instead the sum of the `BLMetalSpecs` concentrations minus the `CATarget`.
  This, paired with a modified set of derivatives for the `MetalComp` in the
  Jacobian matrix, results in Newton-Rapshon changing the metal concentration
  in order to produce chemical conditions that would result in the critical
  accumulation known to cause a toxic effect.
}
