% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getData.R
\name{ReadInputsFromFile}
\alias{ReadInputsFromFile}
\title{Read a BLM Input File}
\usage{
ReadInputsFromFile(
  InputFile,
  NInLab,
  InLabName,
  NInVar,
  InVarName,
  NInComp,
  InCompName
)
}
\arguments{
\item{InputFile}{character; the path and file name to a BLM input file}

\item{NInLab}{integer; number of input label columns}

\item{InLabName}{character vector of length `NInLab`; names of input columns}

\item{NInVar}{integer; Number of input variables}

\item{InVarName}{character vector of length `NInVar`; Names of input
variables}

\item{NInComp}{integer; Number in input components}

\item{InCompName}{character vector of length `NInComp`; names of input
components}
}
\value{
Returns a \code{list} object with the following components:
\describe{
 \item{\code{NObs}}{integer; the number of chemistry observations}
 \item{\code{InLabObs}}{matrix with \code{Obs} rows and \code{InLab}
   columns; the input labels for each observation}
 \item{\code{InVarObs}}{matrix with \code{Obs} rows and \code{InVar}
   columns; the input variables for each observation}
 \item{\code{InCompObs}}{matrix with \code{Obs} rows and \code{InComp}
   columns; the input component concentrations for each observation}
}
}
\description{
`ReadInputsFromFile` will read a BLM input file, assuming it matches the
problem as defined by the input arguments.
}
\examples{
## Not Run
# thisProblem = DefineProblem("my_parameter_file.dat")
# do.call(
#   "ReadInputsFromFile",
#   args = c(thisProblem[names(thisProblem) \%in\%
#              formalArgs("ReadInputsFromFile")],
#            list(InputFile="my_input_file.dat"))
# )
## End Not Run
}
\seealso{
Other BLMEngine Functions: 
\code{\link{CommonParameterDefinitions}}
}
\concept{BLMEngine Functions}
