% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/WriteParamFile.R
\name{WriteParamFile}
\alias{WriteParamFile}
\title{Write a BLM Parameter File}
\usage{
WriteParamFile(ThisProblem, ParamFile)
}
\arguments{
\item{ThisProblem}{A list object with a structure like that returned by
`BlankProblem()`.}

\item{ParamFile}{a character value, indicating the file path and name of the
parameter file to write.}
}
\value{
(invisibly) TRUE, if successful.
}
\description{
This function will take a BLM chemical problem list object and
  turn it into a parameter file, effectively doing the opposite of
  `DefineProblem`.
}
\examples{
tf = tempfile()
WriteParamFile(ThisProblem = carbonate_system_problem, ParamFile = tf)
DefineProblem(tf)

}
